requirements.txt
Flask>=2.3.3
line-bot-sdk>=3.8.0
aiohttp>=3.9.0
google-auth>=2.23.4
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.108.0
openai>=1.3.7
python-dotenv>=1.0.0
gunicorn>=21.2.0
SQLAlchemy>=2.0.23
gspread>=5.12.0
oauth2client>=4.1.3
schedule>=1.2.0
requests>=2.31.0 